src/app/webApp/requirements.txt
pandas
matplotlib
requests
pyarrow
cryptocompare
dash
dash-bootstrap-components
dash-core-components
dash-extensions
dash-html-components
dash-renderer
dash-table
dask
Flask
Flask-Caching
Flask-Compress
json5
jsonschema
numpydoc
croniter
google-cloud-bigquery[bqstorage,pandas]
mysql-connector-python
configparser
